% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_relatorio.R
\name{baixar_relatorio}
\alias{baixar_relatorio}
\title{Baixa um relatório do Pipefy e salva em um arquivo .xlsx.}
\usage{
baixar_relatorio(id_pipe, id_relatorio, diretorio, headers = c())
}
\arguments{
\item{id_pipe}{ID do pipe.}

\item{id_relatorio}{ID do relatório.}

\item{diretorio}{Diretório onde o relatório será salvo.}

\item{headers}{Headers adicionais para a requisição HTTP.}
}
\value{
NULL.
}
\description{
Baixa um relatório do Pipefy e salva em um arquivo .xlsx.
}
