% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{card}
\alias{card}
\title{Retorna informações de um card no Pipefy.}
\usage{
card(id, response_fields = NULL, headers = c())
}
\arguments{
\item{id}{ID do card.}

\item{response_fields}{Campos a serem retornados pela API.}

\item{headers}{Headers adicionais para a requisição HTTP.}
}
\value{
Lista com informações do card ou NULL se o card não for encontrado.
}
\description{
Retorna informações de um card no Pipefy.
}
