% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criar_card.R
\name{criar_card}
\alias{criar_card}
\title{Cria um novo card em um pipe no Pipefy.}
\usage{
criar_card(
  pipe_id,
  atributos,
  cards_ligados_id = NULL,
  response_fields = NULL,
  headers = c()
)
}
\arguments{
\item{pipe_id}{ID do pipe.}

\item{atributos}{Atributos do card.}

\item{cards_ligados_id}{IDs dos cards ligados.}

\item{response_fields}{Campos a serem retornados pela API.}

\item{headers}{Headers adicionais para a requisição HTTP.}
}
\value{
Lista com informações do card criado ou NULL se o card não for criado.
}
\description{
Cria um novo card em um pipe no Pipefy.
}
