% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listar_ids_cards_fase.R
\name{listar_ids_cards_fase}
\alias{listar_ids_cards_fase}
\title{Lista os IDs dos cards em uma fase do Pipefy.}
\usage{
listar_ids_cards_fase(id_fase, headers = c())
}
\arguments{
\item{id_fase}{ID da fase.}

\item{headers}{Headers adicionais para a requisição HTTP.}
}
\value{
Tibble com os IDs dos cards.
}
\description{
Lista os IDs dos cards em uma fase do Pipefy.
}
