% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listar_membros_organizacao.R
\name{listar_membros_organizacao}
\alias{listar_membros_organizacao}
\title{Lista os membros da organizacao no Pipefy}
\usage{
listar_membros_organizacao(header = c())
}
\arguments{
\item{header}{Headers adicionais para a requisição HTTP.}
}
\value{
Tibble com informações dos membros do pipe.
}
\description{
Lista os membros da organizacao no Pipefy
}
