% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mover_card.R
\name{mover_card}
\alias{mover_card}
\title{Move um card para uma fase diferente no Pipefy.}
\usage{
mover_card(card_id, fase_destino_id, response_fields = NULL, headers = c())
}
\arguments{
\item{card_id}{ID do card.}

\item{fase_destino_id}{ID da fase de destino.}

\item{response_fields}{Campos a serem retornados pela API.}

\item{headers}{Headers adicionais para a requisição HTTP.}
}
\value{
Lista com informações do card movido.
}
\description{
Move um card para uma fase diferente no Pipefy.
}
