% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparar_requisicao.R
\name{preparar_requisicao}
\alias{preparar_requisicao}
\title{prepara o objeto para requisicao no pipefy}
\usage{
preparar_requisicao(mock_server = FALSE)
}
\arguments{
\item{mock_server}{A boolean.}
}
\value{
A list.
}
\description{
prepara o objeto para requisicao no pipefy
}
