% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requisicao.R
\name{requisicao}
\alias{requisicao}
\title{Realiza uma requisição para a API do Pipefy.}
\usage{
requisicao(query, headers = c())
}
\arguments{
\item{query}{Query GraphQL a ser enviada para a API.}

\item{headers}{Headers adicionais para a requisição HTTP.}
}
\value{
Resposta da API do Pipefy.
}
\description{
Realiza uma requisição para a API do Pipefy.
}
