% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparar_requisicao.R
\name{preparar_requisicao}
\alias{preparar_requisicao}
\title{Prepara o objeto para requisição no Pipefy.}
\usage{
preparar_requisicao(mock_server = FALSE)
}
\arguments{
\item{mock_server}{Indica se a requisição deve ser feita para um servidor mockado.}
}
\value{
Lista com informações da requisição.
}
\description{
Prepara o objeto para requisição no Pipefy.
}
